# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ever Works Website Template API",
    "description": "REST API documentation for the Ever Works website template. This template provides a complete foundation for building modern web applications with authentication, user management, favorites, and content management features.",
    "version": "1.0.0",
    "contact": {
      "name": "Ever Works Template",
      "url": "https://demo.ever.works"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://demo.ever.works",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/admin/notifications": {
      "get": {
        "tags": ["Admin - Notifications"],
        "summary": "Get admin notifications",
        "description": "Returns notifications for the authenticated admin user with unread count. Limited to 50 most recent notifications ordered by creation date. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminNotificationsResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "notifications": [
                      {
                        "id": "notif_123",
                        "userId": "admin_456",
                        "type": "item_submission",
                        "title": "New Item Submission",
                        "message": "A new item 'Awesome Tool' has been submitted by john@example.com and requires review.",
                        "data": "{\"itemId\":\"item_789\",\"itemName\":\"Awesome Tool\",\"submittedBy\":\"john@example.com\",\"actionUrl\":\"/admin/items/item_789\"}",
                        "isRead": false,
                        "readAt": null,
                        "createdAt": "2024-01-15T10:30:00.000Z",
                        "updatedAt": "2024-01-15T10:30:00.000Z"
                      },
                      {
                        "id": "notif_124",
                        "userId": "admin_456",
                        "type": "user_registered",
                        "title": "New User Registration",
                        "message": "A new user has registered with email: jane@example.com",
                        "data": "{\"userId\":\"user_890\",\"userEmail\":\"jane@example.com\",\"actionUrl\":\"/admin/users/user_890\"}",
                        "isRead": true,
                        "readAt": "2024-01-15T11:00:00.000Z",
                        "createdAt": "2024-01-15T09:45:00.000Z",
                        "updatedAt": "2024-01-15T11:00:00.000Z"
                      }
                    ],
                    "unreadCount": 1
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Forbidden"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": ["Admin - Notifications"],
        "summary": "Create notification",
        "description": "Creates a new notification for a specific user. Requires admin privileges. Data field is optional and will be JSON stringified if provided.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationRequest"
              },
              "example": {
                "type": "system_alert",
                "title": "System Maintenance",
                "message": "The system will be under maintenance from 2 AM to 4 AM UTC.",
                "userId": "user_123",
                "data": {
                  "maintenanceStart": "2024-01-16T02:00:00.000Z",
                  "maintenanceEnd": "2024-01-16T04:00:00.000Z",
                  "affectedServices": ["api", "dashboard"]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateNotificationResponse"
                },
                "example": {
                  "success": true,
                  "notification": {
                    "id": "notif_125",
                    "userId": "user_123",
                    "type": "system_alert",
                    "title": "System Maintenance",
                    "message": "The system will be under maintenance from 2 AM to 4 AM UTC.",
                    "data": "{\"maintenanceStart\":\"2024-01-16T02:00:00.000Z\",\"maintenanceEnd\":\"2024-01-16T04:00:00.000Z\",\"affectedServices\":[\"api\",\"dashboard\"]}",
                    "isRead": false,
                    "readAt": null,
                    "createdAt": "2024-01-15T12:00:00.000Z",
                    "updatedAt": "2024-01-15T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Missing required fields"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/admin/items/{id}": {
      "get": {
        "tags": ["Admin - Items"],
        "summary": "Get item by ID (Admin)",
        "description": "Returns a specific item by its ID. Data is loaded from Git repository. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the item",
            "example": "awesome-productivity-tool"
          }
        ],
        "responses": {
          "200": {
            "description": "Item retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetItemResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "awesome-productivity-tool",
                    "name": "Awesome Productivity Tool",
                    "slug": "awesome-productivity-tool",
                    "description": "A comprehensive productivity tool that helps teams collaborate effectively and manage their tasks efficiently.",
                    "source_url": "https://github.com/awesome/productivity-tool",
                    "category": ["productivity", "collaboration"],
                    "tags": ["productivity", "collaboration", "task-management"],
                    "featured": true,
                    "icon_url": "https://cdn.example.com/icons/productivity-tool.png",
                    "updated_at": "2024-01-15T10:30:00.123Z",
                    "status": "approved",
                    "submitted_by": "user_123",
                    "submitted_at": "2024-01-10T08:15:00.123Z",
                    "reviewed_by": "admin_456",
                    "reviewed_at": "2024-01-12T14:20:00.123Z",
                    "review_notes": "Great tool, approved for listing"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Unauthorized. Admin access required."
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Item not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "tags": ["Admin - Items"],
        "summary": "Update item (Admin)",
        "description": "Updates an item's properties. All fields except ID are optional and can be updated. Data is stored in Git repository. Validates slug format (lowercase, hyphens only), URL format, and checks for duplicate slugs if slug is being updated. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the item",
            "example": "awesome-productivity-tool"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemRequest"
              },
              "example": {
                "name": "Enhanced Productivity Tool",
                "description": "An even more comprehensive productivity tool with advanced features for team collaboration and task management.",
                "category": ["productivity", "collaboration", "project-management"],
                "tags": ["productivity", "collaboration", "task-management", "project-management"],
                "featured": true,
                "status": "approved"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateItemResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "awesome-productivity-tool",
                    "name": "Enhanced Productivity Tool",
                    "slug": "awesome-productivity-tool",
                    "description": "An even more comprehensive productivity tool with advanced features for team collaboration and task management.",
                    "source_url": "https://github.com/awesome/productivity-tool",
                    "category": ["productivity", "collaboration", "project-management"],
                    "tags": ["productivity", "collaboration", "task-management", "project-management"],
                    "featured": true,
                    "icon_url": "https://cdn.example.com/icons/productivity-tool.png",
                    "updated_at": "2024-01-16T15:45:00.123Z",
                    "status": "approved",
                    "submitted_by": "user_123",
                    "submitted_at": "2024-01-10T08:15:00.123Z",
                    "reviewed_by": "admin_456",
                    "reviewed_at": "2024-01-12T14:20:00.123Z",
                    "review_notes": "Great tool, approved for listing"
                  },
                  "message": "Item updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalid_slug": {
                    "summary": "Invalid slug format",
                    "value": {
                      "success": false,
                      "error": "Slug must contain only lowercase letters, numbers, and hyphens"
                    }
                  },
                  "invalid_url": {
                    "summary": "Invalid URL format",
                    "value": {
                      "success": false,
                      "error": "Invalid source URL format"
                    }
                  },
                  "empty_name": {
                    "summary": "Empty name",
                    "value": {
                      "success": false,
                      "error": "Item name cannot be empty"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Unauthorized. Admin access required."
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Item not found"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - duplicate slug",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Item with slug 'awesome-productivity-tool' already exists"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": ["Admin - Items"],
        "summary": "Delete item (Admin)",
        "description": "Deletes an item from the Git repository. This is a permanent deletion that removes the item completely. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the item",
            "example": "awesome-productivity-tool"
          }
        ],
        "responses": {
          "200": {
            "description": "Item deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteItemResponse"
                },
                "example": {
                  "success": true,
                  "message": "Item deleted successfully"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Unauthorized. Admin access required."
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Item not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Session-based authentication using NextAuth.js"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "description": "Additional error details"
          }
        },
        "required": ["success", "error"]
      },
...
      "AdminNotificationsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/NotificationsData"
          }
        },
        "required": ["success", "data"]
      },
      "NotificationsData": {
        "type": "object",
        "properties": {
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            },
            "description": "Array of notifications (limited to 50 most recent)"
          },
          "unreadCount": {
            "type": "integer",
            "description": "Number of unread notifications",
            "example": 1
          }
        },
        "required": ["notifications", "unreadCount"]
      },
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique notification identifier",
            "example": "notif_123"
          },
          "userId": {
            "type": "string",
            "description": "ID of the user who receives the notification",
            "example": "admin_456"
          },
          "type": {
            "type": "string",
            "description": "Type of notification",
            "example": "item_submission"
          },
          "title": {
            "type": "string",
            "description": "Notification title",
            "example": "New Item Submission"
          },
          "message": {
            "type": "string",
            "description": "Notification message content",
            "example": "A new item 'Awesome Tool' has been submitted by john@example.com and requires review."
          },
          "data": {
            "type": "string",
            "nullable": true,
            "description": "JSON stringified additional data",
            "example": "{\"itemId\":\"item_789\",\"itemName\":\"Awesome Tool\",\"submittedBy\":\"john@example.com\",\"actionUrl\":\"/admin/items/item_789\"}"
          },
          "isRead": {
            "type": "boolean",
            "description": "Whether the notification has been read",
            "example": false
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the notification was read",
            "example": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the notification was created",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the notification was last updated",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": ["id", "userId", "type", "title", "message", "isRead", "createdAt", "updatedAt"]
      },
      "CreateNotificationRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "minLength": 1,
            "description": "Type of notification",
            "example": "system_alert"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "description": "Notification title",
            "example": "System Maintenance"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "description": "Notification message content",
            "example": "The system will be under maintenance from 2 AM to 4 AM UTC."
          },
          "userId": {
            "type": "string",
            "minLength": 1,
            "description": "ID of the user who will receive the notification",
            "example": "user_123"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Additional data (will be JSON stringified)",
            "example": {
              "maintenanceStart": "2024-01-16T02:00:00.000Z",
              "maintenanceEnd": "2024-01-16T04:00:00.000Z",
              "affectedServices": ["api", "dashboard"]
            }
          }
        },
        "required": ["type", "title", "message", "userId"]
      },
      "CreateNotificationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "notification": {
            "$ref": "#/components/schemas/Notification"
          }
        },
        "required": ["success", "notification"]
      },
...
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique item identifier"
          },
          "name": {
            "type": "string",
            "description": "Item name"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly item identifier"
          },
          "description": {
            "type": "string",
            "description": "Item description"
          },
          "source_url": {
            "type": "string",
            "format": "uri",
            "description": "URL to the item's source"
          },
          "category": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "description": "Item category (can be string or array of strings)"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Item tags"
          },
          "featured": {
            "type": "boolean",
            "description": "Whether the item is featured (optional)"
          },
          "icon_url": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "URL to the item's icon (optional)"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the item was last updated"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "pending", "approved", "rejected"],
            "description": "Item status"
          },
          "submitted_by": {
            "type": "string",
            "nullable": true,
            "description": "ID of user who submitted the item (optional)"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the item was submitted (optional)"
          },
          "reviewed_by": {
            "type": "string",
            "nullable": true,
            "description": "ID of user who reviewed the item (optional)"
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the item was reviewed (optional)"
          },
          "review_notes": {
            "type": "string",
            "nullable": true,
            "description": "Review notes (optional)"
          }
        },
        "required": ["id", "name", "slug", "description", "source_url", "category", "tags", "updated_at", "status"]
      },
...
      "GetItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/Item"
          }
        },
        "required": ["success", "data"]
      },
      "UpdateItemRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "description": "Updated item name (optional, 3-100 characters)"
          },
          "slug": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "pattern": "^[a-z0-9-]+$",
            "description": "Updated URL-friendly identifier (optional, lowercase letters, numbers, hyphens only)"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 500,
            "description": "Updated item description (optional, 10-500 characters)"
          },
          "source_url": {
            "type": "string",
            "format": "uri",
            "description": "Updated URL to the item's source (optional, valid URL format)"
          },
          "category": {
            "oneOf": [
              {
                "type": "string",
                "description": "Single category as string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Multiple categories as array"
              }
            ],
            "description": "Updated item categories (optional)"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Updated item tags (optional)"
          },
          "featured": {
            "type": "boolean",
            "description": "Updated featured status (optional)"
          },
          "icon_url": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "Updated icon URL (optional, valid URL format or null)"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "pending", "approved", "rejected"],
            "description": "Updated item status (optional)"
          },
          "review_notes": {
            "type": "string",
            "nullable": true,
            "description": "Updated review notes (optional)"
          }
        }
      },
      "UpdateItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/Item"
          },
          "message": {
            "type": "string",
            "example": "Item updated successfully"
          }
        },
        "required": ["success", "data", "message"]
      },
      "DeleteItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Item deleted successfully"
          }
        },
        "required": ["success", "message"]
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Unauthorized"
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Internal server error"
            }
          }
        }
      }
    }
  }
}